pyproject.toml
[tool.poetry]
name = "edu-sphere"
version = "0.1.0"
description = ""
authors = ["Anton Pashinov <troxin1988@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 119

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
stripe = "^10.12.0"
requests = "^2.32.3"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.1.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
coverage = "^7.6.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
